soporte decimales

+++ ProyectoLenguajes/Evaluador.h
@@ -9,5 +9,5 @@ using namespace std;
 
 bool esOperador(char c);
 int precedencia(char c);
-int evaluarPostfija(const std::string& expresion);
+double evaluarPostfija(const std::string& expresion);
 string infijaAPostfija(const std::string& infija);
